requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
appnope==0.1.3
    # via ipython
asgiref==3.5.2
    # via django
asttokens==2.1.0
    # via stack-data
backcall==0.2.0
    # via ipython
build==0.9.0
    # via pip-tools
click==8.1.3
    # via pip-tools
decorator==5.1.1
    # via ipython
django==4.1.3
    # via
    #   -r requirements.in
    #   django-solo
django-solo==2.0.0
    # via -r requirements.in
django-split-settings==1.2.0
    # via -r requirements.in
executing==1.2.0
    # via stack-data
gunicorn==20.1.0
    # via -r requirements.in
ipython==8.6.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
packaging==21.3
    # via build
parso==0.8.3
    # via jedi
pep517==0.13.0
    # via build
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.3.0
    # via -r requirements.in
pip-tools==6.10.0
    # via -r requirements.in
prompt-toolkit==3.0.32
    # via ipython
psycopg2-binary==2.9.5
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.13.0
    # via ipython
pyparsing==3.0.9
    # via packaging
python-dotenv==0.21.0
    # via -r requirements.in
six==1.16.0
    # via asttokens
sorl-thumbnail==12.9.0
    # via -r requirements.in
sqlparse==0.4.3
    # via django
stack-data==0.6.0
    # via ipython
tomli==2.0.1
    # via
    #   build
    #   pep517
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
